Adds file upload properly
File is now converted to a stream before being passed in the body. File
name is taken from command line, split before the file type, 4 bytes of
random hex are inserted at the end to ensure files won't be overwritten.
Mime-types allows for ContentType to be set in the request meaning the
file will no longer be downloaded by default and has proper encoding.

Command to run the script remains the same:
  'node bin/s3upload.js <file-path>'

Packages added:
  - mime-types