Adds back end file upload functionality
Allows for a POST request to /fileuploads to upload a file to AWS S3
with proper MIMEtype, and a Key of the original file name with 4 bytes
of random hex characters added to the end of the filename but before the
file extension. After the call to s3 completes, the return value is used
to create a corresponding object in the mongo database. The user is
currently hardcoded and needs to be set to pull in dynamically based on
the currently signed in user being stored in store.js. the routes now
use Multer to handle multipart data of the file properly.

Breaks command line code ENTIRELY

Package added:
  - Multer